// src/enums/api/laboratories/lab-properties.enum.ts
/* LaboratoryPropertiesEnum */
export enum LaboratoryPropertiesEnum {
  CEP = "CEP",
  COMPANY_NAME = "Nome/Razão social",
  CNPJ = "CNPJ",
  EMAIL = "E-mail",
  ID = "ID",
  LOCALITY_NAME = "Nome do local",
  MAIN_CONTACT = "Contato principal",
  PHONE = "Telefone",
  PUBLIC_PLACE_COMPLEMENT = "Complemento",
  PUBLIC_PLACE_NAME = "Nome do logradouro",
  PUBLIC_PLACE_NUMBER = "Número do logradouro",
  PUBLIC_PLACE_TYPE = "Tipo do logradouro",
  UF = "UF",
}